+++ src/main/kotlin/calculator/CalculatorApplication.kt
@@ -1,8 +1,8 @@
 package calculator
 
-import calculator.calculator.ui.CalculatorController
-import calculator.view.console.ConsoleInput
-import calculator.view.console.ConsoleOutput
+import calculator.ui.CalculatorController
+import view.console.ConsoleInput
+import view.console.ConsoleOutput
 
 val consoleOutput: ConsoleOutput = ConsoleOutput()
 val consoleInput: ConsoleInput = ConsoleInput()
