+++ src/main/kotlin/calculator/ui/CalculatorController.kt
@@ -1,6 +1,6 @@
-package calculator.calculator.ui
+package calculator.ui
 
-import calculator.calculator.domain.Expression
+import calculator.domain.Expression
 
 class CalculatorController {
     fun run(input: String = "") = Expression(input).getResult()
