+++ src/main/kotlin/racing/RacingCarApplication.kt
@@ -1,0 +1,10 @@
+package racing
+
+import view.console.ConsoleInput
+import view.console.ConsoleOutput
+
+val consoleOutput: ConsoleOutput = ConsoleOutput()
+val consoleInput: ConsoleInput = ConsoleInput()
+
+fun main() {
+}
