+++ src/main/kotlin/racing/domain/RacingCar.kt
@@ -1,0 +1,3 @@
+package racing.domain
+
+class RacingCar
