+++ src/main/kotlin/racing/domain/RacingCars.kt
@@ -1,0 +1,3 @@
+package racing.domain
+
+class RacingCars
