+++ src/main/kotlin/racing/ui/RacingCarController.kt
@@ -1,0 +1,3 @@
+package racing.ui
+
+class RacingCarController
