+++ src/main/kotlin/view/console/ConsoleInput.kt
@@ -1,4 +1,4 @@
-package calculator.view.console
+package view.console
 
 class ConsoleInput {
     fun read(): String = readLine()!!
