+++ src/test/kotlin/calculator/domain/ExpressionTest.kt
@@ -1,6 +1,5 @@
 package calculator.domain
 
-import calculator.calculator.domain.Expression
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.junit.jupiter.api.DisplayName
